% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colornames_org.R
\name{get_color_colornames}
\alias{get_color_colornames}
\title{Get color name from colornames.org API}
\usage{
get_color_colornames(hex, return_name_only = TRUE)
}
\arguments{
\item{hex}{Hex color code}

\item{return_name_only}{if FALSE returns color name only}
}
\description{
Return best color name of an hex color from the colornames API
<https://colornames.org/download/>. If no color name exists, it returns NULL`.
You can give a color name and vote for best color names at:
<https://colornames.org/color/YOUR-HEX-CODE>.
}
\examples{
\dontrun{
get_color_colornames(hex = "#0047AB")
}

}
