% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_pizza.R
\name{get_color_colorpizza}
\alias{get_color_colorpizza}
\title{Get color name from color.pizza API}
\source{
\url{https://github.com/meodai/color-name-api}
}
\usage{
get_color_colorpizza(
  hex,
  list = NULL,
  noduplicates = FALSE,
  return_name_only = TRUE
)
}
\arguments{
\item{hex}{character. Hex color code}

\item{list}{character. An available Color Name List.}

\item{noduplicates}{boolean. if TRUE remove duplicates}

\item{return_name_only}{boolean. if TRUE returns color name only}
}
\description{
Rest API that returns a bunch of color names for a given color-value.
}
\examples{
\dontrun{
get_color_colorpizza(
  hex = c("#3D290C","#1c2f11","#2e3f24"),
  list = "wikipedia",
  noduplicates = TRUE)
}

}
