% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colornames_org.R
\name{get_color_colornames_latest}
\alias{get_color_colornames_latest}
\title{Get 100 latest color submissions from colornames.org API}
\usage{
get_color_colornames_latest(return_name_only = TRUE)
}
\arguments{
\item{return_name_only}{if TRUE returns color name only}
}
\value{
A character vector.
}
\description{
Get 100 latest color submissions from colornames.org API
}
\examples{
\dontrun{
get_color_colornames_latest()
}

}
