% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colornames_org.R
\name{get_color_colornames_random}
\alias{get_color_colornames_random}
\title{Get a random color name from colornames.org API}
\usage{
get_color_colornames_random(return_name_only = TRUE)
}
\arguments{
\item{return_name_only}{if TRUE returns color name only}
}
\value{
A character string.
}
\description{
Get a random color name from colornames.org API
}
\examples{
\dontrun{
get_color_colornames_random()
}

}
