% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecolorapi_com.R
\name{get_color_thecolorapi}
\alias{get_color_thecolorapi}
\title{Get color name from thecolorapi.com API}
\usage{
get_color_thecolorapi(
  hex = NULL,
  rgb = NULL,
  hsl = NULL,
  cmyk = NULL,
  format = "json",
  w = 100,
  named = TRUE,
  return_name_only = TRUE
)
}
\arguments{
\item{hex}{string (optional), Valid hex code.}

\item{rgb}{string (optional), Valid rgb color.}

\item{hsl}{string (optional), Valid hsl color.}

\item{cmyk}{string (optional) Valid cmyk color.}

\item{format}{string (optional). Return results as JSON, SVG or HTML page. Example: "html".}

\item{w}{integer (optional). Height of resulting image, only applicable on SVG format. Example: 350.}

\item{named}{boolean (optional) Whether to print the color names on resulting image, only applicable on SVG format. Example: FALSE.}

\item{return_name_only}{boolean. if TRUE returns color name only}
}
\value{
A character string.
}
\description{
Return available identifying information on the given color.
}
\examples{
\dontrun{
get_color_thecolorapi(hex = "0047AB")
get_color_thecolorapi(rgb = paste(0,71,171, sep = ","))
get_color_thecolorapi(hsl = paste("215","100\%","34\%", sep = ","))
get_color_thecolorapi(cmyk = paste(100,58,0,33, sep = ","))
}

}
