% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thecolorapi_com.R
\name{get_color_thecolorapi_scheme}
\alias{get_color_thecolorapi_scheme}
\title{Get color scheme from thecolorapi.com API}
\usage{
get_color_thecolorapi_scheme(
  hex = NULL,
  rgb = NULL,
  hsl = NULL,
  cmyk = NULL,
  count = 5,
  mode = "monochrome",
  format = "json",
  w = 100,
  named = TRUE,
  return_name_only = TRUE
)
}
\arguments{
\item{hex}{string (optional), Valid hex code.}

\item{rgb}{string (optional), Valid rgb color.}

\item{hsl}{string (optional), Valid hsl color.}

\item{cmyk}{string (optional) Valid cmyk color.}

\item{count}{integer (optional). Number of colors to return. Default: 5 Example: 6.}

\item{mode}{string (optional). Define mode by which to generate the scheme from the seed color. Example: "analogic".}

\item{format}{string (optional). Return results as JSON, SVG or HTML page. Example: "html".}

\item{w}{integer (optional). Height of resulting image, only applicable on SVG format. Example: 350.}

\item{named}{boolean (optional) Whether to print the color names on resulting image, only applicable on SVG format. Example: FALSE.}

\item{return_name_only}{boolean. if FALSE returns color name only}
}
\description{
Return a generated scheme for the provided seed color and optional mode.
Color schemes are multi-color combinations chosen according to color-wheel relationsships.
}
\examples{
\dontrun{
get_color_thecolorapi_scheme(hex = "0047AB", count = 3, mode = "complement")
}

}
